requirements.txt
asgiref==3.8.1
Django==5.0.9
sqlparse==0.5.1
typing_extensions==4.12.2
psycopg2
djangorestframework
requests
pypdf